Add a 'distance_ratio_threshold' for 'nearest_atom' parsing
When parsing with 'pos_kind="nearest_atom"', add a check that
computes the ratio of next-nearest / nearest atom. If it is below
the given threshold, an error is thrown. The purpose of this
check is to detect cases where it is inappropriate to use
'nearest_atom' parsing, because the Wannier centers have shifted
so much that they are no longer close to the original atoms,
and might be mis-identified.