Add exception markers with formatting, verbose flag for CLI
Adds a custom exception class with 'exception markers', that is
formatted as '[<exception_marker_name>] <msg>'. The exception
inherits from click.ClickException, meaning it will be shown as
'Error: [<exception_marker_name>] <msg>' when run through the
CLI. This should make error parsing (e.g. in aiida-tbmodels)
much simpler.
Also adds a '--verbose' option to the CLI, without which no
informational messages are printed. This is done to adhere to the
principle that CLI should be 'silent' by default.