Use the rounded `new_R` values in `change_unit_cell`.
Previously, the `new_R` values were checked to be close to an
integer, but then the conversion to `int` happened directly via
`astype(int)` instead of using the rounded values.

This could lead to incorrect results - but it is not expected to
have been common, because the new unit cell in reduced coordinates
is an integer matrix.