Add an option to change the orbital order when folding.
Add an option to choose between different ways of sorting orbitals
in the `fold_model` method: Either by their index in the given
model (the previous default), or by the occurrence of keys in the
`orbital_labels` input (the new default).

The reason for this new feature is that the sort order can be
unpredictable when it is unclear which orbitals end up inside or
outside of the new unit cell.